+++ core/trino-main/src/test/java/io/trino/operator/TestTopNRankingOperator.java
@@ -306,7 +306,7 @@ public class TestTopNRankingOperator
         for (Page page : result.getOutput()) {
             assertThat(page.getChannelCount()).isEqualTo(2);
             for (int i = 0; i < page.getPositionCount(); i++) {
-                assertThat(page.getBlock(1).getByte(i, 0)).isEqualTo(1);
+                assertThat(page.getBlock(1).getByte(i, 0)).isEqualTo((byte) 1);
                 count++;
             }
         }
